requirements.txt
numpy>=1.24
pandas>=1.5
pyyaml>=6.0
matplotlib>=3.7
astropy>=5.0
s3fs>=2024.6.1
astroquery>=0.4.10
scikit-learn>=1.3
pillow>=10.0
tqdm>=4.66

# Deep learning stack
torch>=2.2
torchvision>=0.17
timm>=0.9
